refactor: move shared metrics to internal util package, move controller specific metrics to controller packages
On-behalf-of: @SAP krzysztof.zagorski@sap.com
