+++ internal/controller/cluster/metrics_test.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Greenhouse contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package metrics_test
+package cluster_test
 
 import (
 	"time"
@@ -12,10 +12,9 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 
-	"github.com/cloudoperators/greenhouse/internal/metrics"
-
 	greenhouseapis "github.com/cloudoperators/greenhouse/api"
 	greenhousev1alpha1 "github.com/cloudoperators/greenhouse/api/v1alpha1"
+	"github.com/cloudoperators/greenhouse/internal/controller/cluster"
 	"github.com/cloudoperators/greenhouse/internal/test"
 )
 
@@ -35,7 +34,7 @@ var _ = Describe("Cluster Metrics", Ordered, func() {
 	It("Should return metrics when reconciler is calling", func() {
 		const clusterName = "test-cluster-a"
 		setup := test.NewTestSetup(test.Ctx, test.K8sClient, "clustermetrics")
-		cluster := &greenhousev1alpha1.Cluster{
+		c := &greenhousev1alpha1.Cluster{
 			ObjectMeta: metav1.ObjectMeta{
 				Name:      clusterName,
 				Namespace: setup.Namespace(),
@@ -49,13 +48,13 @@ var _ = Describe("Cluster Metrics", Ordered, func() {
 			},
 		}
 
-		metrics.UpdateClusterMetrics(cluster)
-		counterAfter := prometheusTest.ToFloat64(metrics.KubernetesVersionsGauge.WithLabelValues(cluster.Name, cluster.Namespace, cluster.Status.KubernetesVersion, "test-owner"))
+		cluster.UpdateClusterMetrics(c)
+		counterAfter := prometheusTest.ToFloat64(cluster.KubernetesVersionsGauge.WithLabelValues(c.Name, c.Namespace, c.Status.KubernetesVersion, "test-owner"))
 		Expect(counterAfter).To(BeEquivalentTo(1))
-		tokenExpiry := prometheusTest.ToFloat64(metrics.SecondsToTokenExpiryGauge.WithLabelValues(cluster.Name, cluster.Namespace, "test-owner"))
+		tokenExpiry := prometheusTest.ToFloat64(cluster.SecondsToTokenExpiryGauge.WithLabelValues(c.Name, c.Namespace, "test-owner"))
 		Expect(tokenExpiry).To(BeNumerically(">=", 595))
 		Expect(tokenExpiry).To(BeNumerically("<=", 600))
-		readyGauge := prometheusTest.ToFloat64(metrics.ClusterReadyGauge.WithLabelValues(cluster.Name, cluster.Namespace, "test-owner"))
+		readyGauge := prometheusTest.ToFloat64(cluster.ClusterReadyGauge.WithLabelValues(c.Name, c.Namespace, "test-owner"))
 		Expect(readyGauge).To(BeEquivalentTo(float64(0)), "clusterReady metric should be present and the cluster should not be ready")
 	})
 })
