+++ internal/controller/teamrbac/metrics.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Greenhouse contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package metrics
+package teamrbac
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
