+++ internal/util/metrics.go
@@ -1,7 +1,7 @@
-// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Greenhouse contributors
+// SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Greenhouse contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package metrics
+package util
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
@@ -9,6 +9,7 @@ import (
 
 	greenhouseapis "github.com/cloudoperators/greenhouse/api"
 	greenhousev1alpha1 "github.com/cloudoperators/greenhouse/api/v1alpha1"
+	"github.com/cloudoperators/greenhouse/internal/lifecycle"
 )
 
 type (
@@ -37,17 +38,18 @@ var (
 			Name: "greenhouse_plugin_reconcile_total",
 		},
 		[]string{"pluginDefinition", "clusterName", "plugin", "namespace", "result", "reason", "owned_by"})
-
-	pluginReady = prometheus.NewGaugeVec(
+	OwnedByLabelMissingGauge = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
-			Name: "greenhouse_plugin_ready",
-			Help: "Indicates whether the plugin is ready",
+			Name: "greenhouse_owned_by_label_missing",
+			Help: "Indicates if the greenhouse.sap/owned-by label is missing or invalid",
 		},
-		[]string{"pluginDefinition", "clusterName", "plugin", "namespace", "owned_by"})
+		[]string{"resource_kind", "resource_name", "organization"},
+	)
 )
 
 func init() {
 	crmetrics.Registry.MustRegister(PluginReconcileTotal)
+	crmetrics.Registry.MustRegister(OwnedByLabelMissingGauge)
 }
 
 func UpdatePluginReconcileTotalMetric(plugin *greenhousev1alpha1.Plugin, result MetricResult, reason MetricReason) {
@@ -63,17 +65,15 @@ func UpdatePluginReconcileTotalMetric(plugin *greenhousev1alpha1.Plugin, result 
 	PluginReconcileTotal.With(pluginReconcileTotalLabels).Inc()
 }
 
-func UpdatePluginReadyMetric(plugin *greenhousev1alpha1.Plugin, ready bool) {
-	pluginReadyLabels := prometheus.Labels{
-		"pluginDefinition": plugin.Spec.PluginDefinition,
-		"clusterName":      plugin.Spec.ClusterName,
-		"plugin":           plugin.Name,
-		"namespace":        plugin.Namespace,
-		"owned_by":         plugin.Labels[greenhouseapis.LabelKeyOwnedBy],
+func UpdateOwnedByLabelMissingMetric(resource lifecycle.RuntimeObject, isOwnerLabelMissing bool) {
+	ownedByLabelMissingLabels := prometheus.Labels{
+		"resource_kind": resource.GetObjectKind().GroupVersionKind().Kind,
+		"resource_name": resource.GetName(),
+		"organization":  resource.GetNamespace(),
 	}
-	if ready {
-		pluginReady.With(pluginReadyLabels).Set(1)
+	if isOwnerLabelMissing {
+		OwnedByLabelMissingGauge.With(ownedByLabelMissingLabels).Set(float64(1))
 	} else {
-		pluginReady.With(pluginReadyLabels).Set(0)
+		OwnedByLabelMissingGauge.With(ownedByLabelMissingLabels).Set(float64(0))
 	}
 }
